refactor(context-menu): rename spec
